+++ src/dal/user/dto/UpdateLocationDto.ts
@@ -1,9 +1,9 @@
 import { IsNotEmpty } from "class-validator"
-import { Long } from "typeorm"
+import { Double, Long } from "typeorm"
 
 export class UpdateLocationDto{
     @IsNotEmpty()
     userId: number
-    lat: Long
-    lon: Long
+    lat: Double
+    lon: Double
 }
